Removes User Creation
User creation in Docker and permissions aren't all the same as in typical Linux. Let's avoid all of those issues for now and just run as root. What could go wrong? Also, this adds a workspace directory that can be easily attached as a volume to the outside environment so users can work on local files.